more updates

+++ objects/test/src/anemometer.h
@@ -6,4 +6,4 @@ void anemo_init();
 /**
  * Returns the rotation frequency of the anemometer in Hz.
  */
-float anemo_get();
+double anemo_get();
