+++ objects/test/src/neoPixelStandard.h
@@ -5,7 +5,7 @@
 #ifndef NAOS_PROJECT_NEOPIXELSTANDARD_H
 #define NAOS_PROJECT_NEOPIXELSTANDARD_H
 
-void neoPixelStandard(float power);
+void neoPixelStandard(double power);
 void neoPixelStandard_setup(uint8_t R, uint8_t G, uint8_t B);
 
 #endif  // NAOS_PROJECT_NEOPIXELSTANDARD_H
